Edited styles on Education, Experience and Certification modal windows
