--- src/frontend/src/components/Main/Favorites/Favorites.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Favourites () {
-  return (
-    <h1>
-      Welcome to the Messages page!
-    </h1>
-  )
-}
-
-export default Favourites
