--- src/frontend/src/components/Main/Profile/styles.js
@@ -1,11 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles'
-
-export default makeStyles(() => ({
-
-  text: {
-    marginTop: '75px',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center'
-  }
-}))
