+++ src/frontend/src/components/ProfilePage/ProfileModal/EditCertificationModal/EditCertificationModal.js
@@ -3,26 +3,24 @@ import { Field, Form, Formik } from 'formik'
 import * as Yup from 'yup'
 import { withStyles } from '@material-ui/core/styles'
 import { TextField } from 'formik-material-ui'
-import SharedButton from '../../../shared/SharedButton/SharedButton'
-import toggleModalAction from '../../../redux/Modal/modalActions'
+import SharedButton from '../../../../shared/SharedButton/SharedButton'
+import toggleModalAction from '../../../../redux/Modal/modalActions'
 import { useDispatch } from 'react-redux'
-import month from '../../../data/month.json'
-import year from '../../../data/year.json'
+import month from '../../../../data/month.json'
+import year from '../../../../data/year.json'
 import Grid from '@material-ui/core/Grid'
-import FormikSelect from '../../../shared/FormComponents/FormikSelect'
+import FormikSelect from '../../../../shared/FormComponents/FormikSelect'
 import MuiDialogContent from '@material-ui/core/DialogContent'
 import MuiDialogActions from '@material-ui/core/DialogActions'
-import { createNewCertificationAction } from '../../../redux/Profile/profileActions'
 import Typography from '@material-ui/core/Typography'
-import styles from './styles'
-import convertYearMonthToLocalDate from '../../../utils/convertYearMonthToLocalDate'
+import styles from '../styles'
+import convertLocalDateToYearMonthObj from '../../../../utils/convertLocalDateToYearMonthObj'
+import { deleteCertificationAction, updateCertificationAction } from '../../../../redux/Profile/profileActions'
+import convertYearMonthToLocalDate from '../../../../utils/convertYearMonthToLocalDate'
 
 const DialogContent = withStyles((theme) => ({
   root: {
     padding: theme.spacing(3)
-  },
-  children: {
-    paddingBottom: 25
   }
 }))(MuiDialogContent)
 
@@ -31,12 +29,12 @@ const DialogActions = withStyles((theme) => ({
     width: '100%',
     margin: 0,
     display: 'flex',
-    justifyContent: 'flex-end',
+    justifyContent: 'space-between',
     padding: theme.spacing(1)
   }
 }))(MuiDialogActions)
 
-const CertificationSchema = Yup.object().shape({
+const FORM_VALIDATION = Yup.object().shape({
   name: Yup.string()
     .min(2, 'Too Short!')
     .max(50, 'Too Long!')
@@ -78,8 +76,6 @@ const CertificationSchema = Yup.object().shape({
           'You should follow chronological order',
           function () {
             const {issueDateMonth, issueDateYear, expirationDateMonth, expirationDateYear} = this.parent
-            console.log(expirationDateYear)
-            if (!expirationDateMonth || !expirationDateYear) return true
             const startDateStr = convertYearMonthToLocalDate(issueDateYear, issueDateMonth)
             const endDateStr = convertYearMonthToLocalDate(expirationDateYear, expirationDateMonth)
             const startDateMillis = Date.parse(startDateStr)
@@ -87,34 +83,45 @@ const CertificationSchema = Yup.object().shape({
             return (endDateMillis - startDateMillis) > 0
           })
     }),
-  credentialId: Yup.string(),
-  credentialUrl: Yup.string().url()
+  credentialID: Yup.string(),
+  credentialUrl: Yup.string()
+    .url()
+
 })
 
-export const AddNewCertificationModal = () => {
+const EditCertificationModal = (props) => {
+  const certification = props.certification
+  const {hasExpiryDate} = certification
   const classes = styles()
   const dispatch = useDispatch()
+  const start = convertLocalDateToYearMonthObj(certification.issueDate)
+  const end = hasExpiryDate ? convertLocalDateToYearMonthObj(certification.expirationDate) : null
+
+  const INITIAL_FORM_STATE = {
+    name: certification.name,
+    issuingOrganization: certification.issuingOrganization,
+    hasExpiryDate: certification.hasExpiryDate,
+    issueDateMonth: start.month,
+    issueDateYear: start.year,
+    expirationDateMonth: end ? end.month : '',
+    expirationDateYear: end ? end.year : '',
+    credentialId: certification.credentialId,
+    credentialUrl: certification.credentialUrl
+  }
 
   return (
     <div>
       <Typography variant="subtitle1" className={classes.title}>
-        Add license or certification
+        Edit certification
       </Typography>
       <Formik
         initialValues={{
-          name: '',
-          issuingOrganization: '',
-          hasExpiryDate: false,
-          issueDateMonth: '',
-          issueDateYear: '',
-          expirationDateMonth: '',
-          expirationDateYear: '',
-          credentialId: '',
-          credentialUrl: ''
+          ...INITIAL_FORM_STATE
         }}
-        validationSchema={CertificationSchema}
+        validationSchema={FORM_VALIDATION}
         onSubmit={values => {
-          dispatch(createNewCertificationAction(values))
+          console.log(values)
+          dispatch(updateCertificationAction(values, certification.id))
           dispatch(toggleModalAction())
         }}
       >
@@ -123,7 +130,7 @@ export const AddNewCertificationModal = () => {
             <DialogContent dividers>
               <Grid container spacing={2}>
                 <Grid item xs={12}>
-                  <Field className={classes.formPadding}
+                  <Field
                     name="name"
                     label="Name"
                     required
@@ -137,7 +144,7 @@ export const AddNewCertificationModal = () => {
                     }}/>
                 </Grid>
                 <Grid item xs={12}>
-                  <Field className={classes.formPadding}
+                  <Field
                     name="issuingOrganization"
                     label="Issuing organization"
                     required
@@ -151,19 +158,19 @@ export const AddNewCertificationModal = () => {
                     }}/>
                 </Grid>
                 <Grid item xs={12}>
-                  <label>
+                  <label className={classes.checkbox}>
                     <Field
                       type="checkbox"
                       name="hasExpiryDate"
                     />
-                    <span>Has expiration date</span>
+                    <span>This credential has expiration date</span>
                   </label>
                 </Grid>
                 <Grid item xs={12}>
                   <span>Issue date</span>
                 </Grid>
                 <Grid item xs={6}>
-                  <FormikSelect className={classes.formPadding}
+                  <FormikSelect
                     size="small"
                     name="issueDateMonth"
                     label="Month"
@@ -171,7 +178,7 @@ export const AddNewCertificationModal = () => {
                   />
                 </Grid>
                 <Grid item xs={6}>
-                  <FormikSelect className={classes.formPadding}
+                  <FormikSelect
                     size="small"
                     name="issueDateYear"
                     label="Year"
@@ -182,7 +189,7 @@ export const AddNewCertificationModal = () => {
                   <span>Expiration date</span>
                 </Grid>
                 <Grid item xs={6}>
-                  <FormikSelect className={classes.formPadding}
+                  <FormikSelect
                     disabled={values.hasExpiryDate === false}
                     size="small"
                     name="expirationDateMonth"
@@ -191,7 +198,7 @@ export const AddNewCertificationModal = () => {
                   />
                 </Grid>
                 <Grid item xs={6}>
-                  <FormikSelect className={classes.formPadding}
+                  <FormikSelect
                     disabled={values.hasExpiryDate === false}
                     size="small"
                     name="expirationDateYear"
@@ -200,7 +207,7 @@ export const AddNewCertificationModal = () => {
                   />
                 </Grid>
                 <Grid item xs={12}>
-                  <Field className={classes.formPadding}
+                  <Field
                     name="credentialId"
                     label="Credential ID"
                     type="text"
@@ -212,7 +219,7 @@ export const AddNewCertificationModal = () => {
                     }}/>
                 </Grid>
                 <Grid item xs={12}>
-                  <Field className={classes.formPadding}
+                  <Field
                     name="credentialUrl"
                     label="Credential URL"
                     type="text"
@@ -223,9 +230,16 @@ export const AddNewCertificationModal = () => {
                       shrink: true
                     }}/>
                 </Grid>
+
               </Grid>
             </DialogContent>
-            <DialogActions>
+            <DialogActions classes='justifyContent'>
+              <SharedButton title="Delete certification" variant="outlined" color="secondary"
+                onClick={() => {
+                  dispatch(deleteCertificationAction(certification.id))
+                  dispatch(toggleModalAction())
+                }}/>
+
               <SharedButton type="submit" title="Save"/>
             </DialogActions>
           </Form>
@@ -234,4 +248,5 @@ export const AddNewCertificationModal = () => {
     </div>
   )
 }
-export default AddNewCertificationModal
+
+export default EditCertificationModal
