+++ src/frontend/src/components/ProfilePage/ProfileModal/EditEducationModal/EditEducationModal.js
@@ -2,20 +2,21 @@ import React from 'react'
 import {withStyles} from '@material-ui/core/styles'
 import MuiDialogContent from '@material-ui/core/DialogContent'
 import MuiDialogActions from '@material-ui/core/DialogActions'
-import styles from './styles'
+import styles from '../styles'
 import Typography from '@material-ui/core/Typography'
 import {Form, Formik} from 'formik'
 import * as Yup from 'yup'
-import FormikTextField from '../../../shared/FormComponents/FormikTextField'
-import FormikSelect from '../../../shared/FormComponents/FormikSelect'
+import FormikTextField from '../../../../shared/FormComponents/FormikTextField'
+import FormikSelect from '../../../../shared/FormComponents/FormikSelect'
 import Grid from '@material-ui/core/Grid'
-import month from '../../../data/month.json'
-import year from '../../../data/year.json'
-import SharedButton from '../../../shared/SharedButton/SharedButton'
-import {useDispatch} from 'react-redux'
-import toggleModalAction from '../../../redux/Modal/modalActions'
-import {createNewEducationAction} from '../../../redux/Profile/profileActions'
-import convertYearMonthToLocalDate from '../../../utils/convertYearMonthToLocalDate'
+import month from '../../../../data/month.json'
+import year from '../../../../data/year.json'
+import SharedButton from '../../../../shared/SharedButton/SharedButton'
+import convertLocalDateToYearMonthObj from '../../../../utils/convertLocalDateToYearMonthObj'
+import { updateEducationAction, deleteEducationAction } from '../../../../redux/Profile/profileActions'
+import toggleModalAction from '../../../../redux/Modal/modalActions'
+import { useDispatch } from 'react-redux'
+import convertYearMonthToLocalDate from '../../../../utils/convertYearMonthToLocalDate'
 
 const DialogContent = withStyles((theme) => ({
   root: {
@@ -27,25 +28,28 @@ const DialogActions = withStyles((theme) => ({
   root: {
     margin: 0,
     display: 'flex',
-    justifyContent: 'flex-end',
+    justifyContent: 'space-between',
     padding: theme.spacing(1)
   }
 }))(MuiDialogActions)
 
-const AddEducationModal = () => {
+const EditEducationModal = (props) => {
   const dispatch = useDispatch()
+  const education = props.education
+  const start = convertLocalDateToYearMonthObj(education.dateStart)
+  const end = convertLocalDateToYearMonthObj(education.dateFinish)
 
   const classes = styles()
   const INITIAL_FORM_STATE = {
-    school: '',
-    degreeReceived: '',
-    fieldOfStudy: '',
-    startMonth: '',
-    startYear: '',
-    endMonth: '',
-    endYear: '',
-    activities: '',
-    description: ''
+    school: education.school,
+    degreeReceived: education.degreeReceived,
+    fieldOfStudy: education.fieldOfStudy,
+    startMonth: start.month,
+    startYear: start.year,
+    endMonth: end.month,
+    endYear: end.year,
+    activities: education.activities,
+    description: education.description
   }
   const FORM_VALIDATION = Yup.object().shape({
     school: Yup.string()
@@ -94,7 +98,7 @@ const AddEducationModal = () => {
   return (
     <div>
       <Typography variant="subtitle1" className={classes.title}>
-       Add education
+                Edit education
       </Typography>
 
       <Grid container>
@@ -104,8 +108,7 @@ const AddEducationModal = () => {
           }}
           validationSchema={FORM_VALIDATION}
           onSubmit={values => {
-            console.log(values)
-            dispatch(createNewEducationAction(values))
+            dispatch(updateEducationAction(values, education.id))
             dispatch(toggleModalAction())
           }}
         >
@@ -145,39 +148,47 @@ const AddEducationModal = () => {
                     placeholder="Ex: Business"
                   />
                 </Grid>
-                <Grid item xs={6}>
-                  <FormikSelect
-                    name="startMonth"
-                    label="Month"
-                    size="small"
-                    options={month}
-                    helperText="Start date"
-                  />
+                <Grid container item spacing={2}>
+                  <Grid xs={12}>
+                    <Typography variant='h6' className={classes.subtitle}>Start date</Typography>
+                  </Grid>
+                  <Grid item xs={6}>
+                    <FormikSelect
+                      name="startMonth"
+                      label="Month"
+                      size="small"
+                      options={month}
+                    />
+                  </Grid>
+                  <Grid item xs={6}>
+                    <FormikSelect
+                      name="startYear"
+                      label="Year"
+                      size="small"
+                      options={year}
+                    />
+                  </Grid>
                 </Grid>
-                <Grid item xs={6}>
-                  <FormikSelect
-                    name="startYear"
-                    label="Year"
-                    size="small"
-                    options={year}
-                  />
-                </Grid>
-                <Grid item xs={6}>
-                  <FormikSelect
-                    name="endMonth"
-                    label="Month"
-                    size="small"
-                    options={month}
-                    helperText="End date"
-                  />
-                </Grid>
-                <Grid item xs={6}>
-                  <FormikSelect
-                    name="endYear"
-                    label="Year"
-                    size="small"
-                    options={year}
-                  />
+                <Grid container item spacing={2}>
+                  <Grid xs={12}>
+                    <Typography variant='h6' className={classes.subtitle}>End date</Typography>
+                  </Grid>
+                  <Grid item xs={6}>
+                    <FormikSelect
+                      name="endMonth"
+                      label="Month"
+                      size="small"
+                      options={month}
+                    />
+                  </Grid>
+                  <Grid item xs={6}>
+                    <FormikSelect
+                      name="endYear"
+                      label="Year"
+                      size="small"
+                      options={year}
+                    />
+                  </Grid>
                 </Grid>
                 <Grid item xs={12}>
                   <FormikTextField
@@ -207,6 +218,12 @@ const AddEducationModal = () => {
               </Grid>
             </DialogContent>
             <DialogActions>
+              <SharedButton title="Delete education" variant="outlined" color="secondary"
+                onClick={() => {
+                  dispatch(deleteEducationAction(education.id))
+                  dispatch(toggleModalAction())
+                }}/>
+
               <SharedButton title="Save" type="submit"/>
             </DialogActions>
           </Form>
@@ -215,4 +232,4 @@ const AddEducationModal = () => {
     </div>
   )
 }
-export default AddEducationModal
+export default EditEducationModal
