+++ src/frontend/src/components/ProfilePage/ProfileModal/EditIntroModal/EditIntroModal.js
@@ -2,15 +2,15 @@ import React from 'react'
 import {withStyles} from '@material-ui/core/styles'
 import MuiDialogContent from '@material-ui/core/DialogContent'
 import MuiDialogActions from '@material-ui/core/DialogActions'
-import styles from './styles'
+import styles from '../styles'
 import Typography from '@material-ui/core/Typography'
 import {Form, Formik} from 'formik'
 import * as Yup from 'yup'
-import FormikTextField from '../../../shared/FormComponents/FormikTextField'
+import FormikTextField from '../../../../shared/FormComponents/FormikTextField'
 import Grid from '@material-ui/core/Grid'
-import SharedButton from '../../../shared/SharedButton/SharedButton'
-import {editIntroAction} from '../../../redux/Profile/profileActions'
-import toggleModalAction from '../../../redux/Modal/modalActions'
+import SharedButton from '../../../../shared/SharedButton/SharedButton'
+import {editIntroAction} from '../../../../redux/Profile/profileActions'
+import toggleModalAction from '../../../../redux/Modal/modalActions'
 import {useDispatch} from 'react-redux'
 
 const DialogContent = withStyles((theme) => ({
